% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{annotate_intersections}
\alias{annotate_intersections}
\title{A function to create tabular output of regions intersected with annotations}
\usage{
annotate_intersections(regions, intersections, use.score = FALSE)
}
\arguments{
\item{regions}{A GenomicRanges object output from \code{read_bed()}.}

\item{intersections}{A list of Hits objects from \code{intersect_annotations()}.}

\item{use.score}{Logical variable. Include the "score" for each genomic region in the tabulated results. Score can mean a variety of things, e.g. percent methylation of a CpG/region or fold-change of a ChIP-seq peak.}
}
\value{
A table with columns from the GenomicRanges object for the regions and corresponding annotations
}
\description{
Given output from \code{intersect_annotations()}, create a table for output which shows peak information and corresponding annotation information.
}
\examples{
bed = system.file('extdata', 'Gm12878_Pol2.narrowPeak.gz', package = 'annotatr')
annotations = c('basic_genes')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = FALSE)

i = intersect_annotations(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE)

t = annotate_intersections(
  regions = d,
  intersections = i,
  use.score = FALSE)

}

