% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr.r
\name{annotate_overlaps}
\alias{annotate_overlaps}
\title{Annotate the overlaps of dataset using a tibble or GRanges object of features}
\usage{
annotate_overlaps(x, annot, verbose = TRUE)
}
\arguments{
\item{x}{a tibble that can be converted to a GRanges object.}

\item{annot}{annotation tibble or GRanges object, with a \code{features} 
of the feature labels.}
}
\description{
Annotate the overlaps of dataset using a tibble or GRanges object of features
}
