% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project.R
\name{at_add_layer}
\alias{at_add_layer}
\title{Add a layer to a project}
\usage{
at_add_layer(project, layer, call = rlang::caller_env())
}
\arguments{
\item{project}{An \link{annot_project}.}

\item{layer}{An \link{annot_layer} with a name not already used in the project.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_project} with the layer added. The input is unchanged.
}
\description{
Add a layer to a project
}
\seealso{
Other projects:
\code{\link[=at_add_roi]{at_add_roi()}},
\code{\link[=at_layers]{at_layers()}},
\code{\link[=at_project]{at_project()}},
\code{\link[=at_remove_layer]{at_remove_layer()}},
\code{\link[=at_remove_roi]{at_remove_roi()}},
\code{\link[=at_rois]{at_rois()}},
\code{\link[=at_summary]{at_summary()}},
\code{\link[=at_validate]{at_validate()}}
}
\concept{projects}
