% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{at_backend_get}
\alias{at_backend_get}
\title{Get a registered backend}
\usage{
at_backend_get(name, call = rlang::caller_env())
}
\arguments{
\item{name}{Single string backend name.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The \code{annot_backend} object.
}
\description{
Get a registered backend
}
\seealso{
Other backends:
\code{\link[=at_backend_detect]{at_backend_detect()}},
\code{\link[=at_backend_list]{at_backend_list()}},
\code{\link[=at_backend_register]{at_backend_register()}},
\code{\link[=at_cubert_export_envi]{at_cubert_export_envi()}},
\code{\link[=at_read_image]{at_read_image()}},
\code{\link[=at_tile]{at_tile()}},
\code{\link[=at_tivita_profile]{at_tivita_profile()}}
}
\concept{backends}
