% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-canvas.R
\name{at_canvas_fit}
\alias{at_canvas_fit}
\title{Zoom the canvas to a bounding box}
\usage{
at_canvas_fit(proxy, bbox = NULL, call = rlang::caller_env())
}
\arguments{
\item{proxy}{An \code{atcanvas_proxy}.}

\item{bbox}{Optional \code{c(xmin, ymin, xmax, ymax)} in level-0 pixels; fits the
whole image when \code{NULL}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The proxy, invisibly.
}
\description{
Zoom the canvas to a bounding box
}
\seealso{
Other shiny:
\code{\link[=atCanvasOutput]{atCanvasOutput()}},
\code{\link[=at_annotate]{at_annotate()}},
\code{\link[=at_app]{at_app()}},
\code{\link[=at_canvas]{at_canvas()}},
\code{\link[=at_canvas_proxy]{at_canvas_proxy()}},
\code{\link[=at_canvas_set_annotations]{at_canvas_set_annotations()}},
\code{\link[=at_canvas_set_band]{at_canvas_set_band()}},
\code{\link[=at_canvas_set_overlay]{at_canvas_set_overlay()}},
\code{\link[=at_canvas_set_selection]{at_canvas_set_selection()}},
\code{\link[=at_canvas_set_tool]{at_canvas_set_tool()}},
\code{\link[=at_tile_source]{at_tile_source()}},
\code{\link[=renderAtCanvas]{renderAtCanvas()}}
}
\concept{shiny}
