% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{at_example_image}
\alias{at_example_image}
\title{Read a bundled example image}
\usage{
at_example_image(
  which = c("tissue", "multiplex", "cube"),
  call = rlang::caller_env()
)
}
\arguments{
\item{which}{One of \code{"tissue"} (RGB raster), \code{"multiplex"} (4-channel
pyramidal TIFF), or \code{"cube"} (40-band ENVI spectral cube).}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_image}.
}
\description{
Read a bundled example image
}
\examples{
\dontshow{if (requireNamespace("magick", quietly = TRUE) || requireNamespace("tiff", quietly = TRUE)) withAutoprint(\{ # examplesIf}
img <- at_example_image("tissue")
at_dims(img)
\dontshow{\}) # examplesIf}
}
\seealso{
Other example data:
\code{\link[=at_example_path]{at_example_path()}},
\code{\link[=at_example_project]{at_example_project()}},
\code{\link[=at_example_session]{at_example_session()}}

Other images:
\code{\link{annotatR-classes}},
\code{\link[=at_bands]{at_bands()}},
\code{\link[=at_dims]{at_dims()}},
\code{\link[=at_hsi_meta]{at_hsi_meta()}},
\code{\link[=at_is_pyramidal]{at_is_pyramidal()}},
\code{\link[=at_is_spectral]{at_is_spectral()}},
\code{\link[=at_meta]{at_meta()}},
\code{\link[=at_n_bands]{at_n_bands()}},
\code{\link[=at_n_levels]{at_n_levels()}},
\code{\link[=at_pixel_size]{at_pixel_size()}},
\code{\link[=at_read_image]{at_read_image()}},
\code{\link[=at_tile]{at_tile()}},
\code{\link[=at_wavelengths]{at_wavelengths()}}
}
\concept{example data}
\concept{images}
