% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{at_example_path}
\alias{at_example_path}
\title{Path to a bundled example file}
\usage{
at_example_path(
  which = c("tissue", "multiplex", "cube"),
  call = rlang::caller_env()
)
}
\arguments{
\item{which}{One of \code{"tissue"}, \code{"multiplex"}, or \code{"cube"}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A single string file path to the example asset.
}
\description{
Path to a bundled example file
}
\examples{
at_example_path("tissue")
}
\seealso{
Other example data:
\code{\link[=at_example_image]{at_example_image()}},
\code{\link[=at_example_project]{at_example_project()}},
\code{\link[=at_example_session]{at_example_session()}}
}
\concept{example data}
