% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{at_example_session}
\alias{at_example_session}
\title{An example batch session}
\usage{
at_example_session(n = 5L, call = rlang::caller_env())
}
\arguments{
\item{n}{Number of images in the queue. Default \code{5}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_session}.
}
\description{
Copies the tissue example into a temporary directory \code{n} times so the queue
has real files to iterate.
}
\examples{
sess <- at_example_session(3)
at_session_status(sess)
}
\seealso{
Other example data:
\code{\link[=at_example_image]{at_example_image()}},
\code{\link[=at_example_path]{at_example_path()}},
\code{\link[=at_example_project]{at_example_project()}}

Other sessions:
\code{\link[=at_current]{at_current()}},
\code{\link[=at_goto]{at_goto()}},
\code{\link[=at_manifest]{at_manifest()}},
\code{\link[=at_next]{at_next()}},
\code{\link[=at_prev]{at_prev()}},
\code{\link[=at_resume]{at_resume()}},
\code{\link[=at_session]{at_session()}},
\code{\link[=at_session_status]{at_session_status()}},
\code{\link[=at_set_status]{at_set_status()}}
}
\concept{example data}
\concept{sessions}
