% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-session.R
\name{at_goto}
\alias{at_goto}
\title{Jump to a specific image}
\usage{
at_goto(session, i, call = rlang::caller_env())
}
\arguments{
\item{session}{An \link{annot_session}.}

\item{i}{Integer image index (1-based) within the queue.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_session} with the cursor at \code{i}.
}
\description{
Jump to a specific image
}
\seealso{
Other sessions:
\code{\link[=at_current]{at_current()}},
\code{\link[=at_example_session]{at_example_session()}},
\code{\link[=at_manifest]{at_manifest()}},
\code{\link[=at_next]{at_next()}},
\code{\link[=at_prev]{at_prev()}},
\code{\link[=at_resume]{at_resume()}},
\code{\link[=at_session]{at_session()}},
\code{\link[=at_session_status]{at_session_status()}},
\code{\link[=at_set_status]{at_set_status()}}
}
\concept{sessions}
