% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-layer.R
\name{at_layer_add}
\alias{at_layer_add}
\title{Add an ROI to a layer}
\usage{
at_layer_add(layer, roi, call = rlang::caller_env())
}
\arguments{
\item{layer}{An \link{annot_layer}.}

\item{roi}{An \link{annot_roi}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_layer} with the ROI appended. The label vocabulary and
colour mapping are extended if the ROI introduces a new label. A colliding
ROI identifier is re-minted to preserve uniqueness. The input is unchanged.
}
\description{
Add an ROI to a layer
}
\examples{
lyr <- at_layer("tumour", labels = "tumour")
lyr <- at_layer_add(lyr, at_roi_rect(0, 0, 10, 10, label = "tumour"))
at_layer_n(lyr)
}
\seealso{
Other layers:
\code{\link[=at_layer]{at_layer()}},
\code{\link[=at_layer_n]{at_layer_n()}},
\code{\link[=at_layer_remove]{at_layer_remove()}},
\code{\link[=at_layer_rois]{at_layer_rois()}},
\code{\link[=at_style]{at_style()}}
}
\concept{layers}
