% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-layer.R
\name{at_layer_n}
\alias{at_layer_n}
\title{Number of ROIs in a layer}
\usage{
at_layer_n(layer, call = rlang::caller_env())
}
\arguments{
\item{layer}{An \link{annot_layer}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
An integer scalar.
}
\description{
Number of ROIs in a layer
}
\seealso{
Other layers:
\code{\link[=at_layer]{at_layer()}},
\code{\link[=at_layer_add]{at_layer_add()}},
\code{\link[=at_layer_remove]{at_layer_remove()}},
\code{\link[=at_layer_rois]{at_layer_rois()}},
\code{\link[=at_style]{at_style()}}
}
\concept{layers}
