% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-layer.R
\name{at_layer_remove}
\alias{at_layer_remove}
\title{Remove an ROI from a layer}
\usage{
at_layer_remove(layer, id, call = rlang::caller_env())
}
\arguments{
\item{layer}{An \link{annot_layer}.}

\item{id}{Single string ROI identifier.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_layer} without the identified ROI. Warns if the ROI is
not present and returns the layer unchanged.
}
\description{
Remove an ROI from a layer
}
\seealso{
Other layers:
\code{\link[=at_layer]{at_layer()}},
\code{\link[=at_layer_add]{at_layer_add()}},
\code{\link[=at_layer_n]{at_layer_n()}},
\code{\link[=at_layer_rois]{at_layer_rois()}},
\code{\link[=at_style]{at_style()}}
}
\concept{layers}
