% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-project.R
\name{at_load_project}
\alias{at_load_project}
\title{Load a project}
\usage{
at_load_project(path, call = rlang::caller_env())
}
\arguments{
\item{path}{Path to a saved project \code{.rds}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The \link{annot_project}. Warns if the file was written by a newer
annotatR.
}
\description{
Load a project
}
\seealso{
Other io:
\code{\link[=at_load_session]{at_load_session()}},
\code{\link[=at_read_geojson]{at_read_geojson()}},
\code{\link[=at_read_qupath]{at_read_qupath()}},
\code{\link[=at_read_rois_csv]{at_read_rois_csv()}},
\code{\link[=at_roi_from_geojson]{at_roi_from_geojson()}},
\code{\link[=at_save_project]{at_save_project()}},
\code{\link[=at_save_session]{at_save_session()}},
\code{\link[=at_write_geojson]{at_write_geojson()}},
\code{\link[=at_write_qupath]{at_write_qupath()}},
\code{\link[=at_write_rois_csv]{at_write_rois_csv()}}
}
\concept{io}
