% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project.R
\name{at_remove_roi}
\alias{at_remove_roi}
\title{Remove an ROI from a project}
\usage{
at_remove_roi(project, id, call = rlang::caller_env())
}
\arguments{
\item{project}{An \link{annot_project}.}

\item{id}{Single string ROI identifier.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_project} without the identified ROI. Warns and returns
the project unchanged if the identifier is not found.
}
\description{
Remove an ROI from a project
}
\seealso{
Other projects:
\code{\link[=at_add_layer]{at_add_layer()}},
\code{\link[=at_add_roi]{at_add_roi()}},
\code{\link[=at_example_project]{at_example_project()}},
\code{\link[=at_layers]{at_layers()}},
\code{\link[=at_project]{at_project()}},
\code{\link[=at_remove_layer]{at_remove_layer()}},
\code{\link[=at_rois]{at_rois()}},
\code{\link[=at_summary]{at_summary()}},
\code{\link[=at_validate]{at_validate()}}
}
\concept{projects}
