% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{at_roi_bbox}
\alias{at_roi_bbox}
\title{ROI bounding box}
\usage{
at_roi_bbox(roi, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A numeric vector \code{c(xmin, ymin, xmax, ymax)} in the ROI's stored
coordinates.
}
\description{
ROI bounding box
}
\examples{
at_roi_bbox(at_roi_circle(50, 50, 10, label = "a"))
}
\seealso{
Other rois:
\code{\link[=at_roi_area]{at_roi_area()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_centroid]{at_roi_centroid()}},
\code{\link[=at_roi_circle]{at_roi_circle()}},
\code{\link[=at_roi_ellipse]{at_roi_ellipse()}},
\code{\link[=at_roi_freehand]{at_roi_freehand()}},
\code{\link[=at_roi_from_sf]{at_roi_from_sf()}},
\code{\link[=at_roi_point]{at_roi_point()}},
\code{\link[=at_roi_polygon]{at_roi_polygon()}},
\code{\link[=at_roi_rect]{at_roi_rect()}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}}
}
\concept{rois}
