% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{at_roi_contains}
\alias{at_roi_contains}
\title{Test whether points fall inside an ROI}
\usage{
at_roi_contains(roi, x, y, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{x, y}{Numeric coordinate vectors of equal length in the ROI's stored
coordinate level.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A logical vector of length \code{length(x)}.
}
\description{
Test whether points fall inside an ROI
}
\examples{
at_roi_contains(at_roi_rect(0, 0, 10, 10, label = "a"), c(5, 20), c(5, 20))
}
\seealso{
Other geometry:
\code{\link[=at_check_containment]{at_check_containment()}},
\code{\link[=at_check_geometry]{at_check_geometry()}},
\code{\link[=at_clamp]{at_clamp()}},
\code{\link[=at_fix_geometry]{at_fix_geometry()}},
\code{\link[=at_flip_y]{at_flip_y()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_distance]{at_roi_distance()}},
\code{\link[=at_roi_overlaps]{at_roi_overlaps()}},
\code{\link[=at_roi_rescale]{at_roi_rescale()}},
\code{\link[=at_roi_ring]{at_roi_ring()}},
\code{\link{at_roi_setops}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}},
\code{\link[=at_rois_overlap]{at_rois_overlap()}},
\code{\link[=at_snap]{at_snap()}},
\code{\link[=at_transform]{at_transform()}}
}
\concept{geometry}
