% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-session.R
\name{at_session_save}
\alias{at_session_save}
\title{Save a session}
\usage{
at_session_save(session, path = NULL, call = rlang::caller_env())
}
\arguments{
\item{session}{An \link{annot_session}.}

\item{path}{Optional output path. Defaults to \verb{_session.rds} in the session's
\code{out_dir}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The session, invisibly.
}
\description{
Save a session
}
\seealso{
Other sessions:
\code{\link[=at_current]{at_current()}},
\code{\link[=at_example_session]{at_example_session()}},
\code{\link[=at_goto]{at_goto()}},
\code{\link[=at_manifest]{at_manifest()}},
\code{\link[=at_next]{at_next()}},
\code{\link[=at_prev]{at_prev()}},
\code{\link[=at_resume]{at_resume()}},
\code{\link[=at_session]{at_session()}},
\code{\link[=at_session_load]{at_session_load()}},
\code{\link[=at_session_status]{at_session_status()}},
\code{\link[=at_set_status]{at_set_status()}}
}
\concept{sessions}
