% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-image.R
\name{at_wavelengths}
\alias{at_wavelengths}
\title{Band centre wavelengths}
\usage{
at_wavelengths(x, call = rlang::caller_env())
}
\arguments{
\item{x}{An \link{annot_image}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A numeric vector of length \code{at_n_bands(x)} for a spectral image, or
\code{numeric(0)} when the image is not spectral.
}
\description{
Band centre wavelengths
}
\seealso{
\code{\link[=at_is_spectral]{at_is_spectral()}}

Other images:
\code{\link{annotatR-classes}},
\code{\link[=at_bands]{at_bands()}},
\code{\link[=at_dims]{at_dims()}},
\code{\link[=at_example_image]{at_example_image()}},
\code{\link[=at_hsi_meta]{at_hsi_meta()}},
\code{\link[=at_is_pyramidal]{at_is_pyramidal()}},
\code{\link[=at_is_spectral]{at_is_spectral()}},
\code{\link[=at_meta]{at_meta()}},
\code{\link[=at_n_bands]{at_n_bands()}},
\code{\link[=at_n_levels]{at_n_levels()}},
\code{\link[=at_pixel_size]{at_pixel_size()}},
\code{\link[=at_read_image]{at_read_image()}},
\code{\link[=at_tile]{at_tile()}}
}
\concept{images}
