% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annotations.R
\name{build_gene_annots}
\alias{build_gene_annots}
\title{A helper function to build genic annotations.}
\usage{
build_gene_annots(
  genome = annotatr::builtin_genomes(),
  annotations = annotatr::builtin_annotations(),
  cache = TRUE
)
}
\arguments{
\item{genome}{The genome assembly.}

\item{annotations}{A character vector with entries of the form \code{[genome]_genes_{1to5kb,promoters,5UTRs,cds,exons,firstexons,introns,intronexonboundaries,exonintronboundaries,3UTRs,intergenic}}.}

\item{cache}{A logical stating whether to use the cache on disk for downloads.}
}
\value{
A list of \code{GRanges} objects with unique \code{id} of the form \code{[type]:i}, \code{tx_id} being the UCSC knownGene transcript name, \code{gene_id} being the Entrez Gene ID, \code{symbol} being the gene symbol from the Entrez ID to symbol mapping in \code{org.db} for that species, and \code{type} being the annotation type.
}
\description{
Using the \code{TxDb.*} group of packages, construct genic annotations consisting of any combination of 1-5kb upstream of a TSS, promoters (< 1kb from TSS), 5UTRs, CDS, exons, first exons, introns, intron/exon and exon/intron boundaries, 3UTRs, and intergenic. For genomes without a \code{TxDb.*} package (e.g. sheep, \code{oviariramb2}), an Ensembl \code{EnsDb} from \code{AnnotationHub} is used instead, and \code{tx_id} and \code{gene_id} are Ensembl IDs.
}
