% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annotations.R
\name{build_mane_txdb}
\alias{build_mane_txdb}
\title{Function to build a TxDb of the MANE Select transcripts}
\usage{
build_mane_txdb(cache = TRUE)
}
\arguments{
\item{cache}{A logical stating whether to use the cache on disk for downloads.}
}
\value{
A \code{TxDb} object.
}
\description{
The transcripts come from the Ensembl version of the MANE GTF, and their gene IDs are replaced by Entrez IDs from the MANE summary.
}
