% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{builtin_genomes}
\alias{builtin_genomes}
\title{Function returning supported TxDb.* genomes}
\usage{
builtin_genomes()
}
\value{
A character vector of genomes for supported TxDb.* packages
}
\description{
Function returning supported TxDb.* genomes
}
\examples{
builtin_genomes()

}
