% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_annotations}
\alias{expand_annotations}
\title{Function to expand annotation shortcuts}
\usage{
expand_annotations(annotations)
}
\arguments{
\item{annotations}{A character vector of annotations, possibly using the shortcut accessors}
}
\value{
A vector of data accession-ized names that are ordered from upstream to downstream in the case of knownGenes and islands to interCGI in the case of cpgs.
}
\description{
Function to expand annotation shortcuts
}

