% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Function to get the directory of the cache}
\usage{
get_cache_dir()
}
\value{
A string giving the cache directory, from the \code{annotatr.cache} option or \code{tools::R_user_dir()}.
}
\description{
Function to get the directory of the cache
}
