% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_rids}
\alias{get_cache_rids}
\title{Function to get the resource IDs of cache entries}
\usage{
get_cache_rids(bfc, rname, exact = TRUE)
}
\arguments{
\item{bfc}{A \code{BiocFileCache} object.}

\item{rname}{A string giving the resource name.}

\item{exact}{A logical stating whether to match \code{rname} exactly (TRUE) or as a prefix (FALSE).}
}
\value{
A character vector of resource IDs.
}
\description{
Function to get the resource IDs of cache entries
}
