% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_chrom_aliases}
\alias{get_chrom_aliases}
\title{Function to map any chromosome alias of a genome to its UCSC-style name}
\usage{
get_chrom_aliases(genome, cache = TRUE)
}
\arguments{
\item{genome}{A string giving the genome assembly, e.g. \code{'mm39'} or \code{'oviariramb2'}.}

\item{cache}{A logical stating whether to use the cache on disk for downloads.}
}
\value{
A named character vector whose names are aliases (e.g. Ensembl, GenBank, RefSeq, and UCSC names) and whose values are UCSC-style names.
}
\description{
Uses the chromAlias file of the UCSC GenArk hub (for \code{GENARK} genomes) or of the UCSC database (\code{goldenPath/[genome]/bigZips}), e.g. to rename the Ensembl sequence names of an \code{EnsDb} (\code{1}, \code{MT}, \code{KI270728.1}) to UCSC names (\code{chr1}, \code{chrM}, \code{chr1_KI270728v1_random}).
}
