% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genark_seqinfo}
\alias{get_genark_seqinfo}
\title{Function to get the Seqinfo of a GenArk genome with UCSC-style names}
\usage{
get_genark_seqinfo(genome, cache = TRUE)
}
\arguments{
\item{genome}{A string giving the genome assembly, one of \code{GENARK$genome}.}

\item{cache}{A logical stating whether to use the cache on disk for downloads.}
}
\value{
A \code{Seqinfo} object.
}
\description{
Function to get the Seqinfo of a GenArk genome with UCSC-style names
}
