% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genark_url}
\alias{get_genark_url}
\title{Function to get the URL of a file in the UCSC GenArk hub for a genome}
\usage{
get_genark_url(genome, file)
}
\arguments{
\item{genome}{A string giving the genome assembly, one of \code{GENARK$genome}.}

\item{file}{A string giving the path of the file within the hub directory, e.g. \code{'GCF_016772045.1.chromAlias.txt'}.}
}
\value{
A string giving the URL.
}
\description{
Function to get the URL of a file in the UCSC GenArk hub for a genome
}
