% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annotations.R
\name{get_mane_summary}
\alias{get_mane_summary}
\title{Function to get the MANE Select summary}
\usage{
get_mane_summary(cache = TRUE)
}
\arguments{
\item{cache}{A logical stating whether to use the cache on disk for downloads.}
}
\value{
A \code{data.frame} with one row per MANE Select transcript, and columns \code{tx_id} (Ensembl transcript ID with version), \code{gene_id} (Entrez ID), and \code{symbol}.
}
\description{
Function to get the MANE Select summary
}
