% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_orgdb_name}
\alias{get_orgdb_name}
\title{Function to get correct org.* package name based on genome}
\usage{
get_orgdb_name(genome = supported_genomes())
}
\arguments{
\item{genome}{A string giving the genome assembly.}
}
\value{
A string giving the name of the correct TxDb.* package name based on \code{genome}.
}
\description{
Function to get correct org.* package name based on genome
}

