% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{hg19_cpg_annotations}
\alias{hg19_cpg_annotations}
\title{hg19_cpg_annotations}
\usage{
data(hg19_cpg_annotations)
}
\description{
A GenomicRanges object for hg19_cpg_annotations annotations
}
\keyword{datasets}

