% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{hg19_enhancers_fantom}
\alias{hg19_enhancers_fantom}
\title{hg19_enhancers_fantom}
\format{A GenomicRanges object.}
\usage{
data(hg19_enhancers_fantom)
}
\description{
A GenomicRanges object for hg19_enhancers_fantom annotations.
}
\details{
Permissive enhancers are taken from the FANTOM5 consortium (http://enhancer.binf.ku.dk/presets/). Description of methods that result in the enhancers can be found in the paper corresponding paper (http://www.ncbi.nlm.nih.gov/pubmed/24670763).
}
\keyword{datasets}

