% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{hg19_promoters_knownGenes}
\alias{hg19_promoters_knownGenes}
\title{hg19_promoters_knownGenes}
\usage{
data(hg19_promoters_knownGenes)
}
\description{
A GenomicRanges object for hg19_promoters_knownGenes annotations
}
\keyword{datasets}

