% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_genomes_doc.R
\name{hg38_chrom_sizes}
\alias{hg38_chrom_sizes}
\title{hg38_chrom_sizes}
\format{A numeric vector with names of chromsomes and values of length of chromsomes.}
\usage{
data(hg38_chrom_sizes)
}
\description{
Chromosome sizes for hg38
}
\details{
Chromosome sizes come from UCSC Genome Browser.
}
\keyword{datasets}

