% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{hg38_cpg_annotations}
\alias{hg38_cpg_annotations}
\title{hg38_cpg_annotations}
\usage{
data(hg38_cpg_annotations)
}
\description{
A GenomicRanges object for hg38_cpg_annotations annotations
}
\keyword{datasets}

