% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{hg38_cpg_inter}
\alias{hg38_cpg_inter}
\title{hg38_cpg_inter}
\format{A GenomicRanges object.}
\usage{
data(hg38_cpg_inter)
}
\description{
A GenomicRanges object for hg38_cpg_inter annotations.
}
\details{
CpG islands are taken from the eponymous track in the Regulation group from the corresponding organism in the UCSC Table Browser. CpG shores are defined as flanking 2Kb sequence upstream/downstream of the island boundaries. CpG shelves are defined as flanking 2Kb further upstream/downstream from the CpG shores. InterCGI is the remaining genomic sequence.
}
\keyword{datasets}

