% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{hg38_introns3UTRs_knownGenes}
\alias{hg38_introns3UTRs_knownGenes}
\title{hg38_introns3UTRs_knownGenes}
\usage{
data(hg38_introns3UTRs_knownGenes)
}
\description{
A GenomicRanges object for hg38_introns3UTRs_knownGenes annotations
}
\keyword{datasets}

