% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{load_cached_annotation}
\alias{load_cached_annotation}
\title{Function to load a built annotation from the cache}
\usage{
load_cached_annotation(code, genome)
}
\arguments{
\item{code}{A string giving the annotation code.}

\item{genome}{A string giving the genome assembly.}
}
\value{
A \code{GRanges} object, or \code{NULL} if the annotation isn't cached.
}
\description{
A cached annotation that can't be read is removed so it will be rebuilt.
}
