% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{mm10_5UTRs_knownGenes}
\alias{mm10_5UTRs_knownGenes}
\title{mm10_5UTRs_knownGenes}
\usage{
data(mm10_5UTRs_knownGenes)
}
\description{
A GenomicRanges object for mm10_5UTRs_knownGenes annotations
}
\keyword{datasets}

