% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_genomes_doc.R
\name{mm10_chrom_sizes}
\alias{mm10_chrom_sizes}
\title{mm10_chrom_sizes}
\format{A numeric vector with names of chromsomes and values of length of chromsomes.}
\usage{
data(mm10_chrom_sizes)
}
\description{
Chromosome sizes for mm10
}
\details{
Chromosome sizes come from UCSC Genome Browser.
}
\keyword{datasets}

