% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{mm10_cpg_annotations}
\alias{mm10_cpg_annotations}
\title{mm10_cpg_annotations}
\usage{
data(mm10_cpg_annotations)
}
\description{
A GenomicRanges object for mm10_cpg_annotations annotations
}
\keyword{datasets}

