% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{mm10_promoters_knownGenes}
\alias{mm10_promoters_knownGenes}
\title{mm10_promoters_knownGenes}
\usage{
data(mm10_promoters_knownGenes)
}
\description{
A GenomicRanges object for mm10_promoters_knownGenes annotations
}
\keyword{datasets}

