% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{mm9_introns_knownGenes}
\alias{mm9_introns_knownGenes}
\title{mm9_introns_knownGenes}
\usage{
data(mm9_introns_knownGenes)
}
\description{
A GenomicRanges object for mm9_introns_knownGenes annotations
}
\keyword{datasets}

