% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{mm9_knownGenes_5UTRs}
\alias{mm9_knownGenes_5UTRs}
\title{mm9_knownGenes_5UTRs}
\format{A GenomicRanges object.}
\usage{
data(mm9_knownGenes_5UTRs)
}
\description{
A GenomicRanges object for mm9_knownGenes_5UTRs annotations. KnownGenes are taken from the Genes group, UCSC Genes track, and knownGenes table of the UCSC Table Browser.
Promoter sequence is defined as <1Kb upstream of a transcription start site (TSS, txStart). 1-5Kb upstream sequence is exactly that, upstream of the promoter sequence. 5-prime UTR sequence is defined as the txStart to the cdsStart. Coding sequence (CDS) is defined as cdsStart to cdsEnd. Exons are defined as the matching intervals of exonStarts to exonEnds. Introns are defined as filling the space beween the exons. 3-prime UTR sequence is defined as cdsEnd to txEnd.
}
\details{
More fine-grained definitions include exons and introns that occur in 5-prime UTR, CDS, and 3-prime UTR sequence. First exons and introns are exactly that, while respecting strandedness.
}
\keyword{datasets}

