% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_summary}
\alias{order_summary}
\title{Function to order a column of a tbl_df or grouped_df}
\usage{
order_summary(summary, col, col_order)
}
\arguments{
\item{summary}{A \code{tbl_df} or \code{grouped_df} result from a \code{summarized} function.}

\item{col}{A string indicating which column of \code{summary} to order}

\item{col_order}{A character vector indicating the order of \code{col}.}
}
\value{
A modified version of \code{summary} with \code{col} turned into a factor with levels ordered by \code{col_order}.
}
\description{
Function to order a column of a tbl_df or grouped_df
}

