% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_df}
\alias{read_df}
\title{Convert a data.frame into a GenomicRanges object}
\usage{
read_df(df, genome, stranded = FALSE, use.score = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} following the BED6 format with possibly additional data columns.}

\item{genome}{Gives the genome assembly (human or mouse). Can be one of 'hg19', 'hg38', 'mm9' or 'mm10' in order to use built-in annotations. Can be any genome if custom annotations are used.}

\item{stranded}{Logical variable. If TRUE, strand attribute of GenomicRanges object is drawn from 6th column of BED file. Default \code{FALSE}}

\item{use.score}{Logical variable. If TRUE, score attribute of GenomicRanges object is drawn from 5th column of BED file. Default \code{FALSE}}
}
\value{
A GenomicRanges object with ranges limited by genome and BED file. The GenomicRanges object is sorted if it is detected to be unsorted, and the regions are unique.
}
\description{
\code{read_bed} converts a \code{data.frame} into a \code{GenomicRanges} object. It checks to see if it follows the appropriate BED6+ (Browser Extensible Data) format, and checks validity of genome. Unlike in the BED specification, the score column can be continuously valued. The names need not be unique.
}
\examples{
# Convert an existing data.frame into GRanges for annotation
df = data.frame(
  chr = c('chr1','chr1','chr1','chr1'),
  start = c(10800, 11000, 27800, 29000),
  end = c(10900, 11100, 28800, 29300),
  name = c('A','A','B','B'),
  score = c(87, 45, 34, 62),
  strand = c('*','*','*','*'),
  stringsAsFactors=FALSE)

gr = read_df(
  df = df,
  genome = 'hg19',
  stranded = FALSE,
  use.score=TRUE)

}

