% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reformat_hmm_codes}
\alias{reformat_hmm_codes}
\title{Function to recode classes from chromHMM type column}
\usage{
reformat_hmm_codes(hmm_codes)
}
\arguments{
\item{hmm_codes}{in the original form from UCSC Genome Browser track.}
}
\value{
A character vector of chromHMM classes with numbers and underscores removed.
}
\description{
Function to recode classes from chromHMM type column
}
