% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_cached_annotation}
\alias{save_cached_annotation}
\title{Function to save a built annotation in the cache}
\usage{
save_cached_annotation(gr, code, genome)
}
\arguments{
\item{gr}{A \code{GRanges} object of the annotation.}

\item{code}{A string giving the annotation code.}

\item{genome}{A string giving the genome assembly.}
}
\value{
The path of the cached file, invisibly.
}
\description{
Any earlier versions of the annotation in the cache are removed.
}
