% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_mcols}
\alias{standardize_mcols}
\title{Function to give an annotation the standard mcols}
\usage{
standardize_mcols(gr)
}
\arguments{
\item{gr}{A \code{GRanges} object of an annotation.}
}
\value{
The \code{GRanges} object with the standard mcols, in order.
}
\description{
Adds any missing columns of \code{id}, \code{tx_id}, \code{gene_id}, \code{symbol}, \code{entrez_id}, \code{ensembl_id}, and \code{type} as \code{NA}, and drops any others, so annotations can be combined with \code{c()}.
}
