% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_id_version}
\alias{strip_id_version}
\title{Function to remove the version from Ensembl IDs}
\usage{
strip_id_version(ids)
}
\arguments{
\item{ids}{A character vector of IDs, e.g. \code{'ENSG00000121410.14'}.}
}
\value{
The IDs without versions, e.g. \code{'ENSG00000121410'}. IDs without a version are unchanged.
}
\description{
Function to remove the version from Ensembl IDs
}
