% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_order_tbl}
\alias{subset_order_tbl}
\title{Function to subset a tbl_df or grouped_df by a column}
\usage{
subset_order_tbl(tbl, col, col_order)
}
\arguments{
\item{tbl}{A \code{tbl_df} or \code{grouped_df}.}

\item{col}{A string indicating which column of of \code{tbl} to subset and order}

\item{col_order}{A character vector indicating the order of \code{col}.}
}
\value{
A modified version of \code{summary} with \code{col} subsetted by \code{col_order}.
}
\description{
Function to subset a tbl_df or grouped_df by a column
}

