% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_summary}
\alias{subset_summary}
\title{Function to subset a tbl_df or grouped_df by a column}
\usage{
subset_summary(summary, col, col_order)
}
\arguments{
\item{summary}{A \code{tbl_df} or \code{grouped_df} result from a \code{summarized} function.}

\item{col}{A string indicating which column of of \code{summary} to subset and/or order}

\item{col_order}{A character vector indicating the order of \code{col}.}
}
\value{
A modified version of \code{summary} with \code{col} subsetted by \code{col_order}.
}
\description{
Function to subset a tbl_df or grouped_df by a column
}

