% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_annotation}
\alias{summarize_annotation}
\title{Summarize annotation counts}
\usage{
summarize_annotation(annotated_regions)
}
\arguments{
\item{annotated_regions}{The \code{tbl_df} result of \code{annotate_regions()}.}
}
\value{
A \code{tbl_df} of the number of regions per annotation.
}
\description{
Given a \code{dplyr::tbl_df} of annotated regions, count the number of regions in each annotation.
}
\examples{
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_annotation(i)

}

