% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_annotations}
\alias{summarize_annotations}
\title{Summarize annotation counts}
\usage{
summarize_annotations(annotated_regions, annotated_random, quiet = FALSE)
}
\arguments{
\item{annotated_regions}{The \code{GRanges} result of \code{annotate_regions()}.}

\item{annotated_random}{The \code{GRanges} result of \code{annotate_regions()} on a background of regions, i.e. the regions your data could have come from (e.g. all tested CpGs when \code{annotated_regions} are differentially methylated CpGs). Despite the name, this should not be randomized regions; \code{randomize_regions()} is deprecated.}

\item{quiet}{Print progress messages (FALSE) or not (TRUE).}
}
\value{
A \code{tbl_df} of the number of regions per annotation type.
}
\description{
Given a \code{GRanges} of annotated regions, count the number of regions in each annotation type. If \code{annotated_random} is not missing, then the same is computed for the background regions, labeled "Background" in the \code{data_type} column.
}
\details{
If a region is annotated to multiple annotations of the same \code{annot.type}, the region will only be counted once. For example, if a region were annotated to multiple exons, it would only count once toward the exons, but if it were annotated to an exon and an intron, it would count towards both.
}
\examples{
   ### An example of differentially methylated (DM) regions compared to
   ### all regions tested for differential methylation

   # Get premade CpG annotations
   data('annotations', package = 'annotatr')

   file = system.file('extdata', 'IDH2mut_v_NBM_multi_data_chr9.txt.gz', package = 'annotatr')
   extraCols = c(diff_meth = 'numeric', mu1 = 'numeric', mu0 = 'numeric')
   r = read_regions(con = file, genome = 'hg19', extraCols = extraCols,
       rename_score = 'pval', rename_name = 'DM_status', format = 'bed')

   # Annotate all tested regions, which are the background
   tested_annots = annotate_regions(
       regions = r,
       annotations = annotations,
       ignore.strand = TRUE,
       quiet = FALSE)

   # The data are the DM regions
   dm_annots = tested_annots[tested_annots$DM_status != 'none']

   # Summarize the annotated DM regions
   s = summarize_annotations(annotated_regions = dm_annots)

   # Summarize the annotated DM regions and the background
   s_bg = summarize_annotations(
       annotated_regions = dm_annots,
       annotated_random = tested_annots)

}
