% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_name}
\alias{summarize_name}
\title{Summarize scores over annotations}
\usage{
summarize_name(annotated_regions)
}
\arguments{
\item{annotated_regions}{The tbl_df result of \code{annotate_intersections()}.}
}
\value{
A tbl_df of the counts of (annot_type, data_name) pairs.
}
\description{
Given a dplyr::tbl_df of annotated regions, summarize names over (annot_type, data_name) pairs. The assumption is that the name column in the input BED file consists of a set of non-unique names effectively serving as a label for the region.
}
\examples{
# An example of differentially methylated regions classified as DM up, DM down, or no DM
bed = system.file('extdata', 'IDH2mut_v_NBM_names_scores_chr9.txt.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = intersect_annotations(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE)

t = annotate_intersections(
  regions = d,
  intersections = i,
  use.score = TRUE)

s = summarize_name(t)

}

