% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_score}
\alias{summarize_score}
\title{Summarize scores over annotations}
\usage{
summarize_score(annotated_regions)
}
\arguments{
\item{annotated_regions}{The \code{tbl_df} result of \code{annotate_regions()}.}
}
\value{
A \code{tbl_df} of the mean of scores over (annot_type, annot_id) pairs.
}
\description{
Given a \code{dplyr::tbl_df} of annotated regions, summarize scores over (annot_type, annot_id) pairs.
}
\examples{
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('hg19_basicgenes','hg19_cpgs')

d = read_bed(file = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_score(i)

}

