% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_annotations}
\alias{supported_annotations}
\title{Function listing which annotations are available.}
\usage{
supported_annotations()
}
\value{
A character vector of available annotations.
}
\description{
This includes the shortcuts. The \code{expand_annotations()} function helps
handle the shortcuts.
}
\examples{
supported_annotations()

}
