% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_genomes}
\alias{supported_genomes}
\title{Function returning supported TxDb.* genomes}
\usage{
supported_genomes()
}
\value{
A character vector of genomes for supported TxDb.* packages
}
\description{
Function returning supported TxDb.* genomes
}
\examples{
supported_genomes()

}

