% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{tabulate_intersections}
\alias{tabulate_intersections}
\title{A function to create tabular output of regions intersected with annotations}
\usage{
tabulate_intersections(regions, intersections)
}
\arguments{
\item{intersections}{A list of Hits objects from \code{intersect_annotations()}.}
}
\value{
A table with columns from the GenomicRanges object for the regions and corresponding annotations
}
\description{
Given output from \code{intersect_annotations()}, create a table for output which shows peak information and corresponding annotation information.
}
\examples{
bed = system.file('extdata', 'Gm12878_Pol2.narrowPeak.gz', package = 'annotatr')
annotations = c('basic_genes')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE)

i = intersect_annotations(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE)

t = tabulate_intersections(
  regions = d,
  intersections = i)

}

