% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_annotations}
\alias{tidy_annotations}
\title{Function to tidy up annotation accessors for visualization}
\usage{
tidy_annotations(annotations)
}
\arguments{
\item{annotations}{A character vector of annotations, in the order they are to appear in the visualization.}
}
\value{
A list of mappings from original annotation names to names ready for visualization.
}
\description{
Function to tidy up annotation accessors for visualization
}

