% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr.r
\name{tidy_gff}
\alias{tidy_gff}
\title{Tidy a GFF tibble}
\usage{
tidy_gff(x, chroms = NULL, features = c("five_prime_UTR", "three_prime_UTR",
  "exon", "intron", "gene", "CDS"))
}
\arguments{
\item{x}{a tibble of a GFF file.}

\item{chroms}{a character vector of chromosomes to include.}

\item{features}{a character vector of features
of the file to include (others filtered out).}
}
\description{
Tidy a GFF tibble
}
