% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ucsc_genark_seqlevels}
\alias{ucsc_genark_seqlevels}
\title{Function to rename the sequences of a GenArk genome's GRanges to UCSC-style names}
\usage{
ucsc_genark_seqlevels(gr, genome, seqinfo = get_genark_seqinfo(genome))
}
\arguments{
\item{gr}{A \code{GRanges} or \code{GRangesList} with sequence names that are any alias in the GenArk chromAlias file.}

\item{genome}{A string giving the genome assembly, one of \code{GENARK$genome}.}

\item{seqinfo}{A \code{Seqinfo} object from \code{get_genark_seqinfo()}.}
}
\value{
\code{gr} with UCSC-style sequence names and the full \code{seqinfo}.
}
\description{
Function to rename the sequences of a GenArk genome's GRanges to UCSC-style names
}
