% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_categorical}
\alias{visualize_categorical}
\title{Visualize names over annotations}
\usage{
visualize_categorical(summarized_cats, x, fill = NULL, x_order = NULL,
  fill_order = NULL, position = "stack", plot_title = NULL,
  legend_title = NULL, x_label = NULL, y_label = NULL)
}
\arguments{
\item{summarized_cats}{The \code{grouped_df} result of \code{summarize_categorical()}.}

\item{x}{One of 'annot_type' or 'data_name', indicating whether annotation classes or data classes will appear on the x-axis.}

\item{fill}{One of 'annot_type', 'data_name', or \code{NULL}, indicating whether annotation classes or data classes will fill the bars. If \code{NULL} then the bars will be the total counts of the x classes.}

\item{x_order}{A character vector that subsets and orders the x classes. Default \code{NULL}, uses existing values.}

\item{fill_order}{A character vector that subsets and orders the fill classes. Default \code{NULL}, uses existing values.}

\item{position}{A string which has the same possible values as in \code{ggplot2::geom_bar(..., position)}, i.e., 'stack', 'fill', 'dodge', etc.}

\item{plot_title}{A string used for the title of the plot. Default \code{NULL}, no title displayed.}

\item{legend_title}{A string used for the legend title to describe fills (if fill is not \code{NULL}). Default \code{NULL}, displays corresponding variable name.}

\item{x_label}{A string used for the x-axis label. Default \code{NULL}, corresponding variable name used.}

\item{y_label}{A string used for the y-axis label. Default \code{NULL}, corresponding variable name used.}
}
\value{
A \code{ggplot} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Given a \code{dplyr::grouped_df} of name aggregated annotations (from \code{summarize_categorical()}), visualize the the distribution of \code{fill} in \code{x}. A bar representing the distribution of all \code{fill} in \code{x} will be added according to the contents of \code{fill}. This serves as a background to compare against.
}
\examples{
dm = system.file('extdata', 'IDH2mut_v_NBM_multi_data_chr9.txt.gz', package = 'annotatr')
annotations = c('hg19_basicgenes','hg19_cpgs')

dm_d = read_bed(
  file = dm,
  col.names=c('chr','start','end','DM_status','pval','strand','diff_meth','mu1','mu0'),
  genome = 'hg19',
  stranded = FALSE,
  use.score = TRUE)

dm_r = annotate_regions(
  regions = dm_d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

dm_sc = summarize_categorical(
  annotated_regions = dm_r,
  by = c('annot_type', 'DM_status'))

dm_order = c(
  'hyper',
  'hypo',
  'none')
genes_order = c(
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')

dm_vn = visualize_categorical(
  summarized_cats = dm_sc,
  x = 'DM_status',
  fill = 'annot_type',
  x_order = dm_order,
  fill_order = genes_order,
  position = 'fill',
  legend_title = 'knownGene Annotations',
  x_label = 'DM status',
  y_label = 'Proportion')

}

