% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_coannotations}
\alias{visualize_coannotations}
\title{Visualize pair-wise annotations across regions}
\usage{
visualize_coannotations(annotated_regions, annotation_order = NULL,
  plot_title = NULL, axes_label = NULL)
}
\arguments{
\item{annotated_regions}{The \code{tbl_df} result of \code{annotate_regions()}.}

\item{annotation_order}{A character vector which doubles as the subset of annotations desired for visualization as well as the ordering. If \code{NULL}, all annotations are displayed.}

\item{plot_title}{A string used for the title of the plot. Default \code{NULL}, no title is displayed.}

\item{axes_label}{A string used for the axis labels. Default \code{NULL}, corresponding variable name used.}
}
\value{
A \code{ggplot} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.

dm = system.file('extdata', 'IDH2mut_v_NBM_multi_data_chr9.txt.gz', package = 'annotatr')
annotations = c('hg19_basicgenes','hg19_cpgs','hg19_enhancers_fantom')

dm_d = read_bed(
  file = dm,
  col.names=c('chr','start','end','DM_status','pval','strand','diff_meth','mu1','mu0'),
  genome = 'hg19',
  stranded = FALSE,
  use.score = TRUE)

dm_r = annotate_regions(
  regions = dm_d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

all_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter',
  'hg19_enhancers_fantom',
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')

chip_vcas = visualize_coannotations(
  annotated_regions = dm_r,
  annotation_order = all_order,
  axes_label = 'Annotations',
  plot_title = 'Co-occurrence of Annotations')
}
\description{
All co-occurring annotations associated with a region are computed and displayed as a heatmap.
}

