% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_score}
\alias{visualize_score}
\title{Visualize score distributions over annotations}
\usage{
visualize_score(summarized_scores, annotation_order)
}
\arguments{
\item{summarized_scores}{The \code{grouped_df} result of \code{summarize_score()}.}

\item{annotation_order}{A character vector which doubles as the subset of annotations desired for visualization as well as the ordering.}
}
\value{
A \code{ggplot} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Given a \code{dplyr::grouped_df} of score aggregated annotations (from \code{summarize_score()}), visualize the distribution of scores over the annotation types.
}
\examples{
########################################################################
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_score(i)

cpgs_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter')
v_cpgs = visualize_score(s, cpgs_order)

genes_order = c(
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')
v_genes = visualize_score(s, genes_order)

########################################################################
# An example of percent methylation at CpG sites
bed = system.file('extdata', '2607_mc_hmc_perc_meth_chr21.txt.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_score(i)

cpgs_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter')
v_cpgs = visualize_score(s, cpgs_order)

genes_order = c(
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')
v_genes = visualize_score(s, genes_order)

}

