% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr.R
\docType{package}
\name{annotatr-package}
\alias{annotatr}
\alias{annotatr-package}
\title{annotatr: annotations for R}
\description{
Adds support for annotations in R
}
\author{
\strong{Maintainer}: Aviral Goel \email{goel.aviral@gmail.com}

}
