% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatR-package.R
\docType{package}
\name{annotatR-package}
\alias{annotatR}
\alias{annotatR-package}
\title{annotatR: Multi-Layer Region-of-Interest Annotation for Images and Spectral Cubes}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Annotate multiple layers of regions of interest across whole-slide microscopy images, hyperspectral data cubes, and conventional rasters. Annotations are stored as validated simple-feature geometries in image pixel coordinates with explicit pyramid-level transforms, and can be rasterised to binary, labelled, or multi-class integer masks for downstream segmentation and classification workflows. A batch annotation application built on 'shiny' and 'OpenSeadragon' supports resumable sessions across image queues, live mask preview, and export to 'GeoJSON', 'QuPath', and 'TIFF' formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTTIR/annotatR}
  \item \url{https://cttir.github.io/annotatR/}
  \item Report bugs at \url{https://github.com/CTTIR/annotatR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
