% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-canvas.R
\name{at_canvas_set_band}
\alias{at_canvas_set_band}
\title{Set the displayed band or RGB triple}
\usage{
at_canvas_set_band(proxy, band, call = rlang::caller_env())
}
\arguments{
\item{proxy}{An \code{atcanvas_proxy}.}

\item{band}{A band index or a length-3 vector of band indices.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The proxy, invisibly.
}
\description{
Set the displayed band or RGB triple
}
\seealso{
Other shiny:
\code{\link[=atCanvasOutput]{atCanvasOutput()}},
\code{\link[=at_annotate]{at_annotate()}},
\code{\link[=at_canvas]{at_canvas()}},
\code{\link[=at_canvas_fit]{at_canvas_fit()}},
\code{\link[=at_canvas_proxy]{at_canvas_proxy()}},
\code{\link[=at_canvas_set_annotations]{at_canvas_set_annotations()}},
\code{\link[=at_canvas_set_overlay]{at_canvas_set_overlay()}},
\code{\link[=at_canvas_set_tool]{at_canvas_set_tool()}},
\code{\link[=at_tile_source]{at_tile_source()}},
\code{\link[=renderAtCanvas]{renderAtCanvas()}}
}
\concept{shiny}
