% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-canvas.R
\name{at_canvas_set_overlay}
\alias{at_canvas_set_overlay}
\title{Overlay a mask on the canvas}
\usage{
at_canvas_set_overlay(proxy, mask, alpha = 0.5, call = rlang::caller_env())
}
\arguments{
\item{proxy}{An \code{atcanvas_proxy}.}

\item{mask}{An \code{annot_mask}.}

\item{alpha}{Overlay opacity. Default \code{0.5}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The proxy, invisibly.
}
\description{
Overlay a mask on the canvas
}
\seealso{
Other shiny:
\code{\link[=atCanvasOutput]{atCanvasOutput()}},
\code{\link[=at_annotate]{at_annotate()}},
\code{\link[=at_canvas]{at_canvas()}},
\code{\link[=at_canvas_fit]{at_canvas_fit()}},
\code{\link[=at_canvas_proxy]{at_canvas_proxy()}},
\code{\link[=at_canvas_set_annotations]{at_canvas_set_annotations()}},
\code{\link[=at_canvas_set_band]{at_canvas_set_band()}},
\code{\link[=at_canvas_set_tool]{at_canvas_set_tool()}},
\code{\link[=at_tile_source]{at_tile_source()}},
\code{\link[=renderAtCanvas]{renderAtCanvas()}}
}
\concept{shiny}
