% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{at_clamp}
\alias{at_clamp}
\title{Clip an ROI to image bounds}
\usage{
at_clamp(roi, dims, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{dims}{Integer \code{c(width, height)} bounds in pixels.}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_roi} clipped to \verb{[0, width] x [0, height]}. Warns if the
clip empties the geometry.
}
\description{
Clip an ROI to image bounds
}
\examples{
at_clamp(at_roi_rect(-5, -5, 20, 20, label = "a"), dims = c(10, 10))
}
\seealso{
Other geometry:
\code{\link[=at_check_containment]{at_check_containment()}},
\code{\link[=at_check_geometry]{at_check_geometry()}},
\code{\link[=at_fix_geometry]{at_fix_geometry()}},
\code{\link[=at_flip_y]{at_flip_y()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_contains]{at_roi_contains()}},
\code{\link[=at_roi_distance]{at_roi_distance()}},
\code{\link[=at_roi_overlaps]{at_roi_overlaps()}},
\code{\link[=at_roi_rescale]{at_roi_rescale()}},
\code{\link[=at_roi_ring]{at_roi_ring()}},
\code{\link{at_roi_setops}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}},
\code{\link[=at_rois_overlap]{at_rois_overlap()}},
\code{\link[=at_snap]{at_snap()}},
\code{\link[=at_transform]{at_transform()}}
}
\concept{geometry}
