% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-layer.R
\name{at_layer_rois}
\alias{at_layer_rois}
\title{Layer ROI table}
\usage{
at_layer_rois(layer, image = NULL, call = rlang::caller_env())
}
\arguments{
\item{layer}{An \link{annot_layer}.}

\item{image}{Optional \link{annot_image} for accurate level-0 coordinate transforms.}

\item{call}{The calling environment, for error reporting.}
}
\value{
An ROI table (see \code{\link[=at_rois]{at_rois()}} for the column contract). A 0-row table
with the same columns when the layer has no ROIs.
}
\description{
Layer ROI table
}
\seealso{
Other layers:
\code{\link[=at_layer]{at_layer()}},
\code{\link[=at_layer_add]{at_layer_add()}},
\code{\link[=at_layer_n]{at_layer_n()}},
\code{\link[=at_layer_remove]{at_layer_remove()}},
\code{\link[=at_style]{at_style()}}
}
\concept{layers}
