% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{at_mask_boundary}
\alias{at_mask_boundary}
\title{Object boundaries of a mask}
\usage{
at_mask_boundary(mask, width = 1L, call = rlang::caller_env())
}
\arguments{
\item{mask}{An \code{annot_mask}.}

\item{width}{Integer boundary thickness in pixels. Default \code{1}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A logical matrix marking boundary pixels (where a 4-neighbour has a
different value).
}
\description{
Object boundaries of a mask
}
\seealso{
Other masks:
\code{\link[=at_mask]{at_mask()}},
\code{\link[=at_mask_legend]{at_mask_legend()}},
\code{\link[=at_mask_preview]{at_mask_preview()}},
\code{\link[=at_mask_stack]{at_mask_stack()}},
\code{\link[=at_mask_stats]{at_mask_stats()}},
\code{\link[=at_read_mask]{at_read_mask()}},
\code{\link[=at_write_mask]{at_write_mask()}},
\code{\link[=at_write_masks]{at_write_masks()}},
\code{\link[=plot.annot_mask]{plot.annot_mask()}}
}
\concept{masks}
