% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{at_mask_legend}
\alias{at_mask_legend}
\title{Legend of a mask}
\usage{
at_mask_legend(mask, call = rlang::caller_env())
}
\arguments{
\item{mask}{An \code{annot_mask}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The mask's legend \link[tibble:tibble]{tibble::tibble} (\code{value}, \code{label}, \code{layer},
\code{roi_id}, \code{n_px}, \code{colour}).
}
\description{
Legend of a mask
}
\examples{
at_mask_legend(at_mask(at_example_project(), "labelled"))
}
\seealso{
Other masks:
\code{\link[=at_mask]{at_mask()}},
\code{\link[=at_mask_boundary]{at_mask_boundary()}},
\code{\link[=at_mask_preview]{at_mask_preview()}},
\code{\link[=at_mask_stack]{at_mask_stack()}},
\code{\link[=at_mask_stats]{at_mask_stats()}},
\code{\link[=at_read_mask]{at_read_mask()}},
\code{\link[=at_write_mask]{at_write_mask()}},
\code{\link[=at_write_masks]{at_write_masks()}}
}
\concept{masks}
