% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-image.R
\name{at_n_levels}
\alias{at_n_levels}
\title{Number of pyramid levels}
\usage{
at_n_levels(x, call = rlang::caller_env())
}
\arguments{
\item{x}{An \link{annot_image}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
An integer scalar; \code{1} when the image has no pyramid.
}
\description{
Number of pyramid levels
}
\seealso{
Other images:
\code{\link{annotatR-classes}},
\code{\link[=at_bands]{at_bands()}},
\code{\link[=at_dims]{at_dims()}},
\code{\link[=at_example_image]{at_example_image()}},
\code{\link[=at_is_pyramidal]{at_is_pyramidal()}},
\code{\link[=at_is_spectral]{at_is_spectral()}},
\code{\link[=at_meta]{at_meta()}},
\code{\link[=at_n_bands]{at_n_bands()}},
\code{\link[=at_pixel_size]{at_pixel_size()}},
\code{\link[=at_read_image]{at_read_image()}},
\code{\link[=at_tile]{at_tile()}},
\code{\link[=at_wavelengths]{at_wavelengths()}}
}
\concept{images}
