% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-image.R
\name{at_pixel_size}
\alias{at_pixel_size}
\title{Physical pixel size}
\usage{
at_pixel_size(x, call = rlang::caller_env())
}
\arguments{
\item{x}{An \link{annot_image}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A numeric vector of length 2, the physical size of one level-0 pixel
as \code{c(x, y)}, in the units given by \code{at_meta(x, "pixel_unit")}.
}
\description{
Physical pixel size
}
\seealso{
Other images:
\code{\link{annotatR-classes}},
\code{\link[=at_bands]{at_bands()}},
\code{\link[=at_dims]{at_dims()}},
\code{\link[=at_example_image]{at_example_image()}},
\code{\link[=at_is_pyramidal]{at_is_pyramidal()}},
\code{\link[=at_is_spectral]{at_is_spectral()}},
\code{\link[=at_meta]{at_meta()}},
\code{\link[=at_n_bands]{at_n_bands()}},
\code{\link[=at_n_levels]{at_n_levels()}},
\code{\link[=at_read_image]{at_read_image()}},
\code{\link[=at_tile]{at_tile()}},
\code{\link[=at_wavelengths]{at_wavelengths()}}
}
\concept{images}
