% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{at_plot}
\alias{at_plot}
\title{Plot generic for annotatR objects}
\usage{
at_plot(x, ...)
}
\arguments{
\item{x}{An \link{annot_image}, \link{annot_project}, or \code{annot_mask}.}

\item{...}{Passed to the specific plot function.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot generic for annotatR objects
}
\seealso{
Other plots:
\code{\link[=at_plot_image]{at_plot_image()}},
\code{\link[=at_plot_mask]{at_plot_mask()}},
\code{\link[=at_plot_overlay]{at_plot_overlay()}},
\code{\link[=at_plot_project]{at_plot_project()}},
\code{\link[=at_plot_spectrum]{at_plot_spectrum()}},
\code{\link[=at_plot_summary]{at_plot_summary()}}
}
\concept{plots}
