% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{at_plot_mask}
\alias{at_plot_mask}
\title{Plot a mask}
\usage{
at_plot_mask(mask, legend = TRUE, call = rlang::caller_env())
}
\arguments{
\item{mask}{An \code{annot_mask}.}

\item{legend}{Logical; show the fill legend. Default \code{TRUE}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot a mask
}
\seealso{
Other plots:
\code{\link[=at_plot]{at_plot()}},
\code{\link[=at_plot_image]{at_plot_image()}},
\code{\link[=at_plot_overlay]{at_plot_overlay()}},
\code{\link[=at_plot_project]{at_plot_project()}},
\code{\link[=at_plot_spectrum]{at_plot_spectrum()}},
\code{\link[=at_plot_summary]{at_plot_summary()}}
}
\concept{plots}
