% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{at_plot_summary}
\alias{at_plot_summary}
\title{Plot a project summary}
\usage{
at_plot_summary(project, call = rlang::caller_env())
}
\arguments{
\item{project}{An \link{annot_project}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object of per-label ROI area distributions.
}
\description{
Plot a project summary
}
\seealso{
Other plots:
\code{\link[=at_plot]{at_plot()}},
\code{\link[=at_plot_image]{at_plot_image()}},
\code{\link[=at_plot_mask]{at_plot_mask()}},
\code{\link[=at_plot_overlay]{at_plot_overlay()}},
\code{\link[=at_plot_project]{at_plot_project()}},
\code{\link[=at_plot_spectrum]{at_plot_spectrum()}}
}
\concept{plots}
