% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-session.R
\name{at_prev}
\alias{at_prev}
\title{Step back to the previous image}
\usage{
at_prev(session, call = rlang::caller_env())
}
\arguments{
\item{session}{An \link{annot_session}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_session} with the cursor decremented (clamped at 1).
}
\description{
Step back to the previous image
}
\seealso{
Other sessions:
\code{\link[=at_current]{at_current()}},
\code{\link[=at_goto]{at_goto()}},
\code{\link[=at_manifest]{at_manifest()}},
\code{\link[=at_next]{at_next()}},
\code{\link[=at_resume]{at_resume()}},
\code{\link[=at_session]{at_session()}},
\code{\link[=at_session_load]{at_session_load()}},
\code{\link[=at_session_save]{at_session_save()}},
\code{\link[=at_session_status]{at_session_status()}},
\code{\link[=at_set_status]{at_set_status()}}
}
\concept{sessions}
