% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-project.R
\name{at_remove_layer}
\alias{at_remove_layer}
\title{Remove a layer from a project}
\usage{
at_remove_layer(project, name, call = rlang::caller_env())
}
\arguments{
\item{project}{An \link{annot_project}.}

\item{name}{Single string layer name.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A new \link{annot_project} without the named layer. The input is unchanged.
}
\description{
Remove a layer from a project
}
\seealso{
Other projects:
\code{\link[=at_add_layer]{at_add_layer()}},
\code{\link[=at_add_roi]{at_add_roi()}},
\code{\link[=at_layers]{at_layers()}},
\code{\link[=at_project]{at_project()}},
\code{\link[=at_remove_roi]{at_remove_roi()}},
\code{\link[=at_rois]{at_rois()}},
\code{\link[=at_summary]{at_summary()}},
\code{\link[=at_validate]{at_validate()}}
}
\concept{projects}
