% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{at_roi_buffer}
\alias{at_roi_buffer}
\title{Buffer an ROI}
\usage{
at_roi_buffer(roi, dist, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{dist}{Numeric buffer distance in pixels (may be negative to erode).}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_roi} with the buffered geometry.
}
\description{
Buffer an ROI
}
\examples{
at_roi_buffer(at_roi_point(10, 10, label = "a"), dist = 5)
}
\seealso{
Other rois:
\code{\link[=at_roi_area]{at_roi_area()}},
\code{\link[=at_roi_bbox]{at_roi_bbox()}},
\code{\link[=at_roi_centroid]{at_roi_centroid()}},
\code{\link[=at_roi_circle]{at_roi_circle()}},
\code{\link[=at_roi_ellipse]{at_roi_ellipse()}},
\code{\link[=at_roi_freehand]{at_roi_freehand()}},
\code{\link[=at_roi_from_sf]{at_roi_from_sf()}},
\code{\link[=at_roi_point]{at_roi_point()}},
\code{\link[=at_roi_polygon]{at_roi_polygon()}},
\code{\link[=at_roi_rect]{at_roi_rect()}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}},
\code{\link[=plot.annot_roi]{plot.annot_roi()}}
}
\concept{rois}
