% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{at_roi_buffer}
\alias{at_roi_buffer}
\title{Buffer an ROI}
\usage{
at_roi_buffer(roi, dist, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{dist}{Numeric buffer distance in pixels (may be negative to erode).}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_roi} with the buffered geometry.
}
\description{
Buffer an ROI
}
\examples{
at_roi_buffer(at_roi_point(10, 10, label = "a"), dist = 5)
}
\seealso{
Other geometry:
\code{\link[=at_check_containment]{at_check_containment()}},
\code{\link[=at_check_geometry]{at_check_geometry()}},
\code{\link[=at_clamp]{at_clamp()}},
\code{\link[=at_fix_geometry]{at_fix_geometry()}},
\code{\link[=at_flip_y]{at_flip_y()}},
\code{\link[=at_roi_contains]{at_roi_contains()}},
\code{\link[=at_roi_distance]{at_roi_distance()}},
\code{\link[=at_roi_overlaps]{at_roi_overlaps()}},
\code{\link[=at_roi_rescale]{at_roi_rescale()}},
\code{\link[=at_roi_ring]{at_roi_ring()}},
\code{\link{at_roi_setops}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}},
\code{\link[=at_rois_overlap]{at_rois_overlap()}},
\code{\link[=at_snap]{at_snap()}},
\code{\link[=at_transform]{at_transform()}}
}
\concept{geometry}
