% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{at_roi_distance}
\alias{at_roi_distance}
\title{Distance between two ROIs}
\usage{
at_roi_distance(roi1, roi2, call = rlang::caller_env())
}
\arguments{
\item{roi1, roi2}{\link{annot_roi} objects.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A numeric scalar distance in level-0 pixels.
}
\description{
Distance between two ROIs
}
\seealso{
Other geometry:
\code{\link[=at_check_geometry]{at_check_geometry()}},
\code{\link[=at_clamp]{at_clamp()}},
\code{\link[=at_fix_geometry]{at_fix_geometry()}},
\code{\link[=at_flip_y]{at_flip_y()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_contains]{at_roi_contains()}},
\code{\link[=at_roi_overlaps]{at_roi_overlaps()}},
\code{\link[=at_roi_rescale]{at_roi_rescale()}},
\code{\link{at_roi_setops}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}},
\code{\link[=at_rois_overlap]{at_rois_overlap()}},
\code{\link[=at_snap]{at_snap()}},
\code{\link[=at_transform]{at_transform()}}
}
\concept{geometry}
