% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{at_roi_setops}
\alias{at_roi_setops}
\alias{at_roi_union}
\alias{at_roi_intersect}
\alias{at_roi_difference}
\alias{at_roi_symdiff}
\title{Set operations on ROIs}
\usage{
at_roi_union(..., call = rlang::caller_env())

at_roi_intersect(..., call = rlang::caller_env())

at_roi_difference(..., call = rlang::caller_env())

at_roi_symdiff(..., call = rlang::caller_env())
}
\arguments{
\item{...}{Two or more \link{annot_roi}, or a single \link{annot_layer}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
A single \link{annot_roi} with \code{source = "derived"}.
}
\description{
Combine ROIs geometrically. Each accepts several \link{annot_roi} via \code{...}, or a
single \link{annot_layer}. Multi-part results are returned as one ROI with
\code{MULTIPOLYGON} geometry (the return type is always a single \link{annot_roi}).
}
\examples{
a <- at_roi_rect(0, 0, 6, 6, label = "x")
b <- at_roi_rect(4, 4, 10, 10, label = "x")
at_roi_area(at_roi_intersect(a, b))
}
\seealso{
Other geometry:
\code{\link[=at_check_geometry]{at_check_geometry()}},
\code{\link[=at_clamp]{at_clamp()}},
\code{\link[=at_fix_geometry]{at_fix_geometry()}},
\code{\link[=at_flip_y]{at_flip_y()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_contains]{at_roi_contains()}},
\code{\link[=at_roi_distance]{at_roi_distance()}},
\code{\link[=at_roi_overlaps]{at_roi_overlaps()}},
\code{\link[=at_roi_rescale]{at_roi_rescale()}},
\code{\link[=at_roi_ring]{at_roi_ring()}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}},
\code{\link[=at_rois_overlap]{at_rois_overlap()}},
\code{\link[=at_snap]{at_snap()}},
\code{\link[=at_transform]{at_transform()}}
}
\concept{geometry}
