% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{at_roi_simplify}
\alias{at_roi_simplify}
\title{Simplify an ROI}
\usage{
at_roi_simplify(roi, tolerance, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{tolerance}{Non-negative Douglas-Peucker tolerance (\code{dTolerance}).}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_roi} with simplified geometry.
}
\description{
Simplify an ROI
}
\examples{
at_roi_simplify(at_roi_circle(50, 50, 10, n_seg = 128, label = "a"), 1)
}
\seealso{
Other rois:
\code{\link[=at_roi_area]{at_roi_area()}},
\code{\link[=at_roi_bbox]{at_roi_bbox()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_centroid]{at_roi_centroid()}},
\code{\link[=at_roi_circle]{at_roi_circle()}},
\code{\link[=at_roi_ellipse]{at_roi_ellipse()}},
\code{\link[=at_roi_freehand]{at_roi_freehand()}},
\code{\link[=at_roi_from_sf]{at_roi_from_sf()}},
\code{\link[=at_roi_point]{at_roi_point()}},
\code{\link[=at_roi_polygon]{at_roi_polygon()}},
\code{\link[=at_roi_rect]{at_roi_rect()}}
}
\concept{rois}
