% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{at_roi_simplify}
\alias{at_roi_simplify}
\title{Simplify an ROI}
\usage{
at_roi_simplify(roi, tolerance, call = rlang::caller_env())
}
\arguments{
\item{roi}{An \link{annot_roi}.}

\item{tolerance}{Non-negative Douglas-Peucker tolerance (\code{dTolerance}).}

\item{call}{The calling environment, for error reporting.}
}
\value{
An \link{annot_roi} with simplified geometry.
}
\description{
Simplify an ROI
}
\examples{
at_roi_simplify(at_roi_circle(50, 50, 10, n_seg = 128, label = "a"), 1)
}
\seealso{
Other geometry:
\code{\link[=at_check_geometry]{at_check_geometry()}},
\code{\link[=at_clamp]{at_clamp()}},
\code{\link[=at_fix_geometry]{at_fix_geometry()}},
\code{\link[=at_flip_y]{at_flip_y()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_contains]{at_roi_contains()}},
\code{\link[=at_roi_distance]{at_roi_distance()}},
\code{\link[=at_roi_overlaps]{at_roi_overlaps()}},
\code{\link[=at_roi_rescale]{at_roi_rescale()}},
\code{\link{at_roi_setops}},
\code{\link[=at_rois_overlap]{at_rois_overlap()}},
\code{\link[=at_snap]{at_snap()}},
\code{\link[=at_transform]{at_transform()}}
}
\concept{geometry}
