% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-project.R
\name{at_save_project}
\alias{at_save_project}
\title{Save a project}
\usage{
at_save_project(project, path, overwrite = FALSE, call = rlang::caller_env())
}
\arguments{
\item{project}{An \link{annot_project}.}

\item{path}{Output \code{.rds} path.}

\item{overwrite}{Logical; overwrite an existing file. Default \code{FALSE}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The path, invisibly.
}
\description{
Save a project
}
\examples{
\dontshow{if (requireNamespace("magick", quietly = TRUE) || requireNamespace("tiff", quietly = TRUE)) withAutoprint(\{ # examplesIf}
p <- withr::local_tempfile(fileext = ".rds")
at_save_project(at_example_project(), p)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=at_load_project]{at_load_project()}}

Other io:
\code{\link[=at_load_project]{at_load_project()}},
\code{\link[=at_load_session]{at_load_session()}},
\code{\link[=at_read_geojson]{at_read_geojson()}},
\code{\link[=at_read_qupath]{at_read_qupath()}},
\code{\link[=at_read_rois_csv]{at_read_rois_csv()}},
\code{\link[=at_save_session]{at_save_session()}},
\code{\link[=at_write_geojson]{at_write_geojson()}},
\code{\link[=at_write_qupath]{at_write_qupath()}},
\code{\link[=at_write_rois_csv]{at_write_rois_csv()}}
}
\concept{io}
