% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-session.R
\name{at_session_load}
\alias{at_session_load}
\title{Load a session}
\usage{
at_session_load(path, call = rlang::caller_env())
}
\arguments{
\item{path}{Path to a saved \verb{_session.rds} file.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The restored \link{annot_session}.
}
\description{
Load a session
}
\seealso{
Other sessions:
\code{\link[=at_current]{at_current()}},
\code{\link[=at_goto]{at_goto()}},
\code{\link[=at_manifest]{at_manifest()}},
\code{\link[=at_next]{at_next()}},
\code{\link[=at_prev]{at_prev()}},
\code{\link[=at_resume]{at_resume()}},
\code{\link[=at_session]{at_session()}},
\code{\link[=at_session_save]{at_session_save()}},
\code{\link[=at_session_status]{at_session_status()}},
\code{\link[=at_set_status]{at_set_status()}}
}
\concept{sessions}
