% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-session.R
\name{at_session_status}
\alias{at_session_status}
\title{The session manifest}
\usage{
at_session_status(session, call = rlang::caller_env())
}
\arguments{
\item{session}{An \link{annot_session}.}

\item{call}{The calling environment, for error reporting.}
}
\value{
The manifest \link[tibble:tibble]{tibble::tibble} with columns \code{idx}, \code{path}, \code{name},
\code{status}, \code{project_path}, \code{n_rois}, and \code{modified}.
}
\description{
The session manifest
}
\seealso{
Other sessions:
\code{\link[=at_current]{at_current()}},
\code{\link[=at_example_session]{at_example_session()}},
\code{\link[=at_goto]{at_goto()}},
\code{\link[=at_manifest]{at_manifest()}},
\code{\link[=at_next]{at_next()}},
\code{\link[=at_prev]{at_prev()}},
\code{\link[=at_resume]{at_resume()}},
\code{\link[=at_session]{at_session()}},
\code{\link[=at_session_load]{at_session_load()}},
\code{\link[=at_session_save]{at_session_save()}},
\code{\link[=at_set_status]{at_set_status()}}
}
\concept{sessions}
