% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-canvas.R
\name{atCanvasOutput}
\alias{atCanvasOutput}
\title{Shiny output for the annotation canvas}
\usage{
atCanvasOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Output id.}

\item{width, height}{Output dimensions.}
}
\value{
A Shiny output element.
}
\description{
Shiny output for the annotation canvas
}
\seealso{
Other shiny:
\code{\link[=at_annotate]{at_annotate()}},
\code{\link[=at_canvas]{at_canvas()}},
\code{\link[=at_canvas_fit]{at_canvas_fit()}},
\code{\link[=at_canvas_proxy]{at_canvas_proxy()}},
\code{\link[=at_canvas_set_annotations]{at_canvas_set_annotations()}},
\code{\link[=at_canvas_set_band]{at_canvas_set_band()}},
\code{\link[=at_canvas_set_overlay]{at_canvas_set_overlay()}},
\code{\link[=at_canvas_set_tool]{at_canvas_set_tool()}},
\code{\link[=at_tile_source]{at_tile_source()}},
\code{\link[=renderAtCanvas]{renderAtCanvas()}}
}
\concept{shiny}
