% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annotations.R
\name{build_enhancer_annots}
\alias{build_enhancer_annots}
\title{A helper function to build enhancer annotations for hg19 and mm10 from FANTOM5.}
\usage{
build_enhancer_annots(genome = c("hg19", "mm9"))
}
\arguments{
\item{genome}{The genome assembly.}
}
\value{
A \code{GRanges} object.
}
\description{
A helper function to build enhancer annotations for hg19 and mm10 from FANTOM5.
}

