% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annotations.R
\name{build_hmm_annots}
\alias{build_hmm_annots}
\title{A helper function to build chromHMM annotations for hg19 from UCSC Genome Browser.}
\usage{
build_hmm_annots(genome = c("hg19"), annotations = builtin_annotations())
}
\arguments{
\item{genome}{The genome assembly.}

\item{annotations}{A character vector of valid chromatin state annotatin codes.}
}
\value{
A \code{GRanges} object.
}
\description{
A helper function to build chromHMM annotations for hg19 from UCSC Genome Browser.
}
