% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_annotations}
\alias{check_annotations}
\title{Function to check for valid annotations}
\usage{
check_annotations(annotations)
}
\arguments{
\item{annotations}{A character vector of annotations possibly using the shortcuts}
}
\description{
Gives errors if any annotations are not in supported_annotations(), basicgenes and detailedgenes are used, or the genome prefixes are not the same for all annotations.
}

