% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_annotation_rname}
\alias{get_annotation_rname}
\title{Function to get the cache resource name of a built annotation}
\usage{
get_annotation_rname(code, genome)
}
\arguments{
\item{code}{A string giving the annotation code, e.g. \code{'hg19_genes_promoters'}.}

\item{genome}{A string giving the genome assembly.}
}
\value{
A string giving the resource name.
}
\description{
The name includes the annotatr version, and for gene annotations, the versions of the \code{TxDb.*} and \code{org.*.eg.db} packages, the AnnotationHub EnsDb, or MANE, so updating them causes a rebuild.
}
