% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cellline_from_code}
\alias{get_cellline_from_code}
\title{Function to return cell line from chromatin annotation code}
\usage{
get_cellline_from_code(code)
}
\value{
A string of the cell line used in a chromatin annotation code
}
\description{
Function to return cell line from chromatin annotation code
}
