% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cellline_from_shortcut}
\alias{get_cellline_from_shortcut}
\title{Function to return cell line from chromatin annotation shortcut}
\usage{
get_cellline_from_shortcut(shortcut)
}
\value{
A string of the cell line used in a chromatin annotation shortcut
}
\description{
Function to return cell line from chromatin annotation shortcut
}
