% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genark_aliases}
\alias{get_genark_aliases}
\title{Function to map any chromosome alias of a GenArk genome to its UCSC-style name}
\usage{
get_genark_aliases(genome)
}
\arguments{
\item{genome}{A string giving the genome assembly, one of \code{GENARK$genome}.}
}
\value{
A named character vector whose names are aliases (RefSeq, GenBank, NCBI, UCSC) and whose values are UCSC-style names.
}
\description{
Function to map any chromosome alias of a GenArk genome to its UCSC-style name
}
