% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_orgdb_name}
\alias{get_orgdb_name}
\title{Function to get correct org.* package name based on genome}
\usage{
get_orgdb_name(genome = supported_genomes())
}
\arguments{
\item{genome}{A string giving the genome assembly.}
}
\value{
A string giving the correct org for org.db packages. e.g. hg19 -> Hs.
}
\description{
Function to get correct org.* package name based on genome
}
