% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatr_data_doc.R
\name{mm9_cpg_annotations}
\alias{mm9_cpg_annotations}
\title{mm9_cpg_annotations}
\format{A GenomicRanges object.}
\usage{
data(mm9_cpg_annotations)
}
\description{
A GenomicRanges object for mm9_cpg_annotations annotations.
}
\details{
CpG islands are taken from the eponymous track in the Regulation group from the corresponding organism in the UCSC Table Browser. CpG shores are defined as flanking 2Kb sequence upstream/downstream of the island boundaries. CpG shelves are defined as flanking 2Kb further upstream/downstream from the CpG shores. InterCGI is the remaining genomic sequence.
}
\keyword{datasets}

