% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{plot.annot_roi}
\alias{plot.annot_roi}
\title{Plot an ROI}
\usage{
\method{plot}{annot_roi}(x, ...)
}
\arguments{
\item{x}{An \link{annot_roi}.}

\item{...}{Ignored.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object. The plot is not drawn; print it to render.
}
\description{
Plot an ROI
}
\seealso{
Other rois:
\code{\link[=at_roi_area]{at_roi_area()}},
\code{\link[=at_roi_bbox]{at_roi_bbox()}},
\code{\link[=at_roi_buffer]{at_roi_buffer()}},
\code{\link[=at_roi_centroid]{at_roi_centroid()}},
\code{\link[=at_roi_circle]{at_roi_circle()}},
\code{\link[=at_roi_ellipse]{at_roi_ellipse()}},
\code{\link[=at_roi_freehand]{at_roi_freehand()}},
\code{\link[=at_roi_from_sf]{at_roi_from_sf()}},
\code{\link[=at_roi_point]{at_roi_point()}},
\code{\link[=at_roi_polygon]{at_roi_polygon()}},
\code{\link[=at_roi_rect]{at_roi_rect()}},
\code{\link[=at_roi_simplify]{at_roi_simplify()}}
}
\concept{plots}
\concept{rois}
