% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-roi.R
\name{plot.annot_roi}
\alias{plot.annot_roi}
\title{Plot an ROI}
\usage{
\method{plot}{annot_roi}(x, ...)
}
\arguments{
\item{x}{An \link{annot_roi}.}

\item{...}{Ignored.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object. The plot is not drawn; print it to render.
}
\description{
Plot an ROI
}
\keyword{internal}
