% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_annotations}
\alias{read_annotations}
\title{Convert a BED file into annotations}
\usage{
read_annotations(file, genome, annotation_name)
}
\arguments{
\item{file}{Path to the file with annotations. This file can be either BED3, BED4, or BED5. If BED3, the assumption is that the annotations are not stranded and not named; if BED4,}

\item{genome}{Gives the genome assembly. The genome may be one of hg19, hg38, mm9, or mm10, in which case the user can also annotate regions with built-in annotations. Otherwise, the user may only annotate regions with their custom annotations.}

\item{annotation_name}{A description of the types of annotations present in \code{file}.}
}
\value{
A GenomicRanges object with the name \code{[genome]_custom_[annotation_name]} for use in the \code{annotations} parameter in \code{annotate_regions()} along with built-in annotations.

NOTE: The name of the object created by \code{read_annotations()} should have the form [genome]_custom_[annotation_name], where the genome is either supported (hg19, hg38, mm9, mm10) or not, and the annotation_name is the same as given in \code{read_annotations()}.
}
\description{
Annotations are genomic regions given as 3, 4, or 5 columned tab-delimited files (chrom, start, end, name, strand). If the input is 3-columned, the annotations are assumed to be unstranded and unnamed. In this case, the strand is set to * and the names are set to the numbered sequence \code{annotation_name:n}. If the input is 4-columned, the content of the name column is used, and the strand is set to *. If the input is 5-columned, the name and strand columns are used. If the annotations are unstranded, use '*' for the strand column.
}
\details{
If the names for the annotations are not unique, they will be renamed in the form \code{annotation_name:n}.
}
\examples{
file = system.file('extdata', 'test_annotations_5.bed', package='annotatr')
hg19_custom_test = read_annotations(file = file, genome = 'hg19', annotation_name = 'test')

}

