% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_bed}
\alias{read_bed}
\title{Convert a bed file into a GenomicRanges object}
\usage{
read_bed(filename, genome, stranded = FALSE, use.score = FALSE)
}
\arguments{
\item{filename}{Path to the file with data. File must exist and be in correct format.}

\item{genome}{Gives the genome assembly (human or mouse). Must be one of 'hg19', 'hg38', 'mm9' or 'mm10'}

\item{stranded}{Logical variable. If TRUE, strand attribute of GenomicRanges object is drawn from 6th column of BED file.}

\item{use.score}{Logical variable. If TRUE, score attribute of GenomicRanges object is drawn from 5th column of BED file.}
}
\value{
A GenomicRanges object with ranges limited by genome and BED file. The GenomicRanges object is sorted if it is detected to be unsorted, and the regions are unique. The name column (4th) in the BED file is \code{regionName} attribute and the score column (5th) in the BED file is the \code{score} attribute in the returned GenomicRanges object.
}
\description{
\code{read_bed} reads in data from a BED6 file, checks to see if it follows the appropriate BED (Browser Extensible Data) format, and checks validity of genome. If valid, reads file data into a GenomicRanges object. Unlike in the BED specification, the score column can be continuously valued. The names need not be unique.
}
\examples{
file = system.file('extdata', 'K562_Cjun.narrowPeak.gz', package = 'annotatr')
read_bed(filename = file, genome = 'hg19', stranded = FALSE, use.score = FALSE)

}

