% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget-canvas.R
\name{renderAtCanvas}
\alias{renderAtCanvas}
\title{Render an annotation canvas in Shiny}
\usage{
renderAtCanvas(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression producing an \code{\link[=at_canvas]{at_canvas()}}.}

\item{env}{The environment to evaluate \code{expr} in.}

\item{quoted}{Is \code{expr} already quoted?}
}
\value{
A Shiny render function.
}
\description{
Render an annotation canvas in Shiny
}
\seealso{
Other shiny:
\code{\link[=atCanvasOutput]{atCanvasOutput()}},
\code{\link[=at_canvas]{at_canvas()}},
\code{\link[=at_canvas_fit]{at_canvas_fit()}},
\code{\link[=at_canvas_proxy]{at_canvas_proxy()}},
\code{\link[=at_canvas_set_annotations]{at_canvas_set_annotations()}},
\code{\link[=at_canvas_set_band]{at_canvas_set_band()}},
\code{\link[=at_canvas_set_overlay]{at_canvas_set_overlay()}},
\code{\link[=at_canvas_set_tool]{at_canvas_set_tool()}},
\code{\link[=at_tile_source]{at_tile_source()}}
}
\concept{shiny}
