% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_categorical}
\alias{summarize_categorical}
\title{Summarize categoricals over groupings of annotated regions}
\usage{
summarize_categorical(annotated_regions, by = c("annot_type", "annot_id"))
}
\arguments{
\item{annotated_regions}{The \code{tbl_df} result of \code{annotate_regions()}.}

\item{by}{A character vector to group the data by and tally over. Default is \code{c(annot_type, annot_id)}.}
}
\value{
A \code{tbl_df} of the counts of groupings according to the \code{by} vector.
}
\description{
Given a \code{dplyr::tbl_df} of annotated regions, summarize names over (annot_type, name) pairs. The assumption is that the name column in the input BED file consists of a set of non-unique names effectively serving as a label for the region.
}
\examples{
bed = system.file('extdata', 'test_read_multiple_data_head.bed', package = 'annotatr')
annotations = c('hg19_cpg_islands', 'hg19_knownGenes_promoters')

d = read_bed(
  file = bed,
  col.names=TRUE,
  genome = 'hg19',
  stranded = FALSE,
  use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

sc1 = summarize_categorical(
  annotated_regions = i,
  by = c('annot_type', 'name'))
sc2 = summarize_categorical(
  annotated_regions = i,
  by = c('annot_type', 'diff_exp'))
sc3 = summarize_categorical(
  annotated_regions = i,
  by = c('name','diff_exp'))

}

