% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_score}
\alias{summarize_score}
\title{Summarize scores over annotations}
\usage{
summarize_score(annotated_regions)
}
\arguments{
\item{annotated_regions}{The tbl_df result of \code{annotate_intersections()}.}
}
\value{
A tbl_df of the mean of scores over (annot_type, annot_id) pairs.
}
\description{
Given a dplyr::tbl_df of annotated regions, summarize scores over (annot_type, annot_id) pairs.
}
\examples{
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = F, use.score = TRUE)

i = intersect_annotations(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = T)

t = annotate_intersections(
  regions = d,
  intersections = i,
  use.score = TRUE)

s = summarize_score(t)

}

