% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_annotations}
\alias{supported_annotations}
\title{Function to inform the user which annotations are available.}
\usage{
supported_annotations()
}
\value{
A character vector of accessors for use with data() to load annotations.
}
\description{
Function to inform the user which annotations are available.
}
\examples{
supported_annotations()
data(hg19_cpg_islands)
data(hg38_knownGenes_promoters)

}

