% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_annotations}
\alias{supported_annotations}
\title{Function listing which annotations are available.}
\usage{
supported_annotations()
}
\value{
A character vector of available annotations.
}
\description{
This includes the shortcuts, which cannot be used in the data(...) calls. To call a shortcut, need to use \code{expand_annotations()}.
}
\examples{
supported_annotations()
data(hg19_cpg_islands)
data(hg38_knownGenes_promoters)

}

