% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_genomes}
\alias{supported_genomes}
\title{Function to inform the user which genomes are supported}
\usage{
supported_genomes()
}
\value{
A character vector of genomes for which we have annotations
}
\description{
Function to inform the user which genomes are supported
}
\examples{
supported_genomes()

}

