% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_annotation}
\alias{visualize_annotation}
\title{Visualize the number of regions per annotation}
\usage{
visualize_annotation(summarized_annotations, annotation_order = NULL,
  plot_title = NULL, x_label = NULL, y_label = NULL)
}
\arguments{
\item{summarized_annotations}{The \code{tbl_df} result of \code{summarize_annotations()}.}

\item{annotation_order}{A character vector which doubles as the subset of annotations desired for visualization as well as the ordering. If \code{NULL}, all annotations are displayed.}

\item{plot_title}{A string used for the title of the plot. Default \code{NULL}, no title is displayed.}

\item{x_label}{A string used for the x-axis label. Default \code{NULL}, corresponding variable name used.}

\item{y_label}{A string used for the y-axis label. Default \code{NULL}, corresponding variable name used.}
}
\value{
A \code{ggplot} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Given a \code{dplyr::tbl_df} of counts of regions per annotation (from \code{summarize_annotations()}), visualize the counts per annotation as a bar graph.
}
\examples{
########################################################################
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('hg19_basicgenes','hg19_cpgs')

d = read_bed(file = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_annotations(i)

annots_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter',
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')
v_annots = visualize_annotation(s, annots_order)

}

