% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_name}
\alias{visualize_name}
\title{Visualize scores over annotations}
\usage{
visualize_name(summarized_names, x, fill, x_order, fill_order,
  position = "fill")
}
\arguments{
\item{summarized_names}{The \code{grouped_df} result of \code{summarize_name()}.}

\item{x}{One of 'annot_type' or 'data_name', indicating whether annotation classes or data classes will appear on the x-axis.}

\item{fill}{One of 'annot_type', 'data_name', or NULL, indicating whether annotation classes or data classes will fill the bars. If NULL then the bars will be the total counts of the x classes.}

\item{x_order}{A character vector that subsets and orders the x classes.}

\item{fill_order}{A character vector that subsets and orders the fill classes.}

\item{position}{A string which has the same possible values as in \code{ggplot2::geom_bar(..., position)}, i.e., 'stack', 'fill', 'dodge', etc.}
}
\value{
A \code{ggplot} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Given a \code{dplyr::grouped_df} of name aggregated annotations (from \code{summarize_name()}), visualize the the distribution of \code{annot_type} in \code{data_name}.
}
\examples{
# An example of differentially methylated regions classified as DM up, DM down, or no DM
bed = system.file('extdata', 'IDH2mut_v_NBM_names_scores_chr9.txt.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_name(i)

fill_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter')
x_order = c(
  'DMup',
  'DMdown')
v_cpgs_counts_data_annot = visualize_name(summarized_names=s, x='data_name', fill='annot_type',
  x_order = x_order, fill_order = fill_order, position='stack')
v_cpgs_proportions_data_annot = visualize_name(summarized_names=s, x='data_name', fill='annot_type',
  x_order = x_order, fill_order = fill_order, position='fill')
v_cpgs_nofill_data = visualize_name(summarized_names=s, x='data_name', fill=NULL,
  x_order = x_order, fill_order = fill_order, position='stack')

x_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter')
fill_order = c(
  'DMup',
  'DMdown',
  'noDM')
v_cpgs_counts_annot_data = visualize_name(summarized_names=s, x='annot_type', fill='data_name',
  x_order = x_order, fill_order = fill_order, position='stack')
v_cpgs_proportions_annot_data = visualize_name(summarized_names=s, x='annot_type', fill='data_name',
  x_order = x_order, fill_order = fill_order, position='fill')
v_cpgs_nofill_annot = visualize_name(summarized_names=s, x='annot_type', fill=NULL,
  x_order = x_order, fill_order = fill_order, position='stack')

}

