% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_numerical}
\alias{visualize_numerical}
\title{Visualize score distributions over annotations}
\usage{
visualize_numerical(summarized_scores, annotation_order = NULL,
  bin_width = 10, plot_title = NULL, x_label = NULL, y_label = NULL)
}
\arguments{
\item{summarized_scores}{The \code{grouped_df} result of \code{summarize_numerical()}.}

\item{annotation_order}{A character vector which doubles as the subset of annotations desired for visualization as well as the ordering.}

\item{bin_width}{An integer indicating the bin width of the histogram used for score. Default 10. Select something appropriate for the data.}

\item{plot_title}{A string used for the title of the plot. Default \code{NULL}, no title displayed.}

\item{x_label}{A string used for the x-axis label. Default \code{NULL}, corresponding variable name used.}

\item{y_label}{A string used for the y-axis label. Default \code{NULL}, corresponding variable name used.}
}
\value{
A \code{ggplot} object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Given a \code{dplyr::grouped_df} of score aggregated annotations (from \code{summarize_numerical()}), visualize the distribution of scores over the annotation types.
}
\examples{
########################################################################
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('hg19_basicgenes','hg19_cpgs')

d = read_bed(file = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_numerical(i)

hg19_cpgs_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter')
v_hg19_cpgs = visualize_numerical(s, hg19_cpgs_order)

genes_order = c(
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')
v_genes = visualize_numerical(s, genes_order, bin_width=30)

########################################################################
# An example of percent methylation at CpG sites
bed = system.file('extdata', '2607_mc_hmc_perc_meth_chr21.txt.gz', package = 'annotatr')
annotations = c('hg19_basicgenes','hg19_cpgs')

d = read_bed(file = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = annotate_regions(
  regions = d,
  annotations = annotations,
  ignore.strand = TRUE,
  use.score = TRUE)

s = summarize_numerical(i)

hg19_cpgs_order = c(
  'hg19_cpg_islands',
  'hg19_cpg_shores',
  'hg19_cpg_shelves',
  'hg19_cpg_inter')
v_hg19_cpgs = visualize_numerical(s, hg19_cpgs_order)

genes_order = c(
  'hg19_knownGenes_1to5kb',
  'hg19_knownGenes_promoters',
  'hg19_knownGenes_5UTRs',
  'hg19_knownGenes_exons',
  'hg19_knownGenes_introns',
  'hg19_knownGenes_3UTRs')
v_genes = visualize_numerical(s, genes_order, bin_width=5)

}

