% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_score}
\alias{visualize_score}
\title{Visualize score distributions over annotations}
\usage{
visualize_score(summarized_scores, annotation_order)
}
\arguments{
\item{annotation_order}{A character vector which doubles as the subset of annotations desired for visualization as well as the ordering.}

\item{aggregated_scores}{The tbl_df result of \code{summarize_score()}.}
}
\value{
A ggplot object which can be viewed by calling it, or saved with \code{ggplot2::ggsave}.
}
\description{
Given a dplyr::tbl_df of aggregated scores, visualize the distribution of scores over the annotation types.
}
\examples{
########################################################################
# An example of ChIP-seq peaks with signalValue used for score
bed = system.file('extdata', 'Gm12878_Ezh2_sorted_scores.narrowPeak.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = intersect_annotations(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE)

t = annotate_intersections(
  regions = d,
  intersections = i,
  use.score = TRUE)

s = summarize_score(t)

cpgs_order = c('hg19_cpg_islands','hg19_cpg_shores','hg19_cpg_shelves','hg19_cpg_inter')
v_cpgs = visualize_score(s, cpgs_order)

genes_order = c('1to5kb','promoters','5UTRs','exons','introns','3UTRs')
v_genes = visualize_score(s, genes_order)

########################################################################
# An example of percent methylation at CpG sites
bed = system.file('extdata', '2607_mc_hmc_perc_meth_chr21.txt.gz', package = 'annotatr')
annotations = c('basic_genes','cpgs')

d = read_bed(filename = bed, genome = 'hg19', stranded = FALSE, use.score = TRUE)

i = intersect_annotations(
  regions = d,
  annotations = annotations,
  genome = 'hg19',
  ignore.strand = TRUE)

t = annotate_intersections(
  regions = d,
  intersections = i,
  use.score = TRUE)

s = summarize_score(t)

cpgs_order = c('hg19_cpg_islands','hg19_cpg_shores','hg19_cpg_shelves','hg19_cpg_inter')
v_cpgs = visualize_score(s, cpgs_order)

genes_order = c('1to5kb','promoters','5UTRs','exons','introns','3UTRs')
v_genes = visualize_score(s, genes_order)

}

